// frontend/lib/types.ts
export type Listing = {
  id: number;
  url: string;
  address: string;
  price: number;
  sqmPrice: number;
  sqmSize: string;
  monthlyFee: number;
  roomCount: string;
  imageUrl: string;
  hemnetListingId: string;
  datePublished: string;
  locationId: number;
  balcony: string;
  patio: string;
  floor: string;
  elevator: string;
  createdAt: Date;
};